requirements.txt
openai>=1.14
datasets>=2.19
pandas>=2.2
pyyaml>=6.0
matplotlib>=3.8
tqdm>=4.66
rich>=13.7
